shoes/image.c: Prevent segfault when a cached image path cannot be opened.
When shoes_image_downloaded is called with a successful status and a file that cannot be opened, shoes will segfault when trying to read the file. Check whether the file failed to open and throw an error. The crash was observed in windows using the following script:

Shoes.app do
  def crash
    %w{
      http://hackety.org/images/hackety-org-header.png
      http://shoooes.net/images/shoes-splash.png
      http://hackety.org/images/rxvg-2.png
      http://hackety.org/images/hpricot.png
      http://hackety.org/images/a-note-to-pigeon.png
      http://hackety.org/images/pony-magic.png
      http://hackety.org/images/shoes-downloader.png
      http://hackety.org/images/abebooks3.png
      http://hackety.org/images/shoes-packager.png
    }.each do |url|
      image url
    end
  end

  button('crash'){crash}
end

There is an underlying windows issue with threading or maybe with download handling. This patch makes sense either way.