A hacked solution to the Vista flickering issue.
It does seem that Vista has a serious flaw in its compositing engine
when it comes to updating transparent windows. This does indeed
seem to be a counter-intuitive bug, but it has not been fixed in
Vista, even though it does not exist in either XP or Windows 7.

The bug exists due to it drawing transparent windows over those
of higher z-order without first, you know, compositing them like
the WS_EX_COMPOSITED flag should suggest. There may be a subtle but
better way of fixing this problem, but with whyday approaching, this
will have to do.

For information from other people annoyed by similar problems:
http://www.virtualdub.org/blog/pivot/entry.php?id=149

The solution is dirty but simple. Do the compositing ourselves. Turn
off the transparent flag and ensure the paint order is bottom to top.
The WS_EX_COMPOSITED flag does seem to order the paint order correctly
and overlapping windows do not interfere when not transparent.

For each canvas, we have a double buffer. This is how it would be in the
background anyway. Each time we update the canvas, we copy the image to its
parents and to each child. Thus, every backbuffer will contain the image that
will appear on the screen to the user in the region of the canvas after a full
redraw. This will simulate the transparent effect that the slots seem
to assume.

This will work in most cases with just inefficiency over the
current solution. It may cause some slight graphical glitches with native
controls since they are still marked as transparent, although Windows
certainly has issues with unique uses of the native widgets anyway, sigh.
If they are truly a problem, one could remove the WS_EX_TRANSPARENT bit
when shoes_world->os.doublebuffer is TRUE.

The solution is only activated for Windows version 6.0 (Vista) as can be seen
by the version identification shim in the app initialization code. XP and
Windows 7 will work as they have before. An extra boolean has been added to
the application OS structure that is set when Vista is detected and turns on
the manual compositing mode, which may be useful if the problem comes up in
other interesting versions of Windows. An extra field in the slot OS structure
has to do with not being able to reliably get the parent window using the
Windows API, or at least I was having issues with it.

With this fix, hackety-hack runs perfectly fine in Vista SP2 with Aero Basic.
Which brings up another point that one could _not_ use this technique if they
also detected that Aero was not in use.

I think I will go shower now. Ugh.