Switch from #ifdef NEW_RADIO to OLD_RADIO for #6
It's good enough to be the default considering how miserable the old
radio behaviour was. Now we have groups that work and buttons that don't
spit out invalid surface error messages on the console.

It's not completely fixed until I have images with a little white dot
in the blue circle.