FIXME comment on Symbols
// FIXME: Symbols in Shoes are broken in many ways. For example, the
symbol progress is shared amongst types/progress and types/download,
which causes Shoes to crash if defined in either ones. It should also be
noted that Shoes crashes even if progress is defined here below. It
apparently has to be defined on top of ruby{.c,.h}. Image and
TextBlock/Text/TextLink are also heavily affected by this.
* Shoes3/shoes3#333